requirements.txt
pdfplumber
pyttsx3
pdf2image
pytesseract
pillow